==> 2-distribution_shift.Rmd <==
---
title: "Plots"
author: "Eduardo Yuki Yada"
geometry: margin=1cm
output: 
  pdf_document:
    template: latex-template.tex
params:
  outcome_column: readmission_30d
---

```{r, include=FALSE}
knitr::opts_chunk$set(
  warning = F,
  message = F
)
```

# Imports

```{r warning=F, message=F}
library(tidyverse)
library(yaml)

options(dplyr.summarise.inform = FALSE)
```

# Loading data

```{r}
load('dataset/processed_data.RData') 
load('dataset/processed_dictionary.RData') 

columns_list <- yaml.load_file("./auxiliar/columns_list.yaml")

outcome_column <- params$outcome_column

df[columns_list$outcome_columns] <- lapply(df[columns_list$outcome_columns], as.character)
df[columns_list$outcome_columns] <- lapply(df[columns_list$outcome_columns], as.integer)
```

# Plots

```{r warning=F, message=F, fig.show="hold", out.width="50%", fig.height = 5}
library(lubridate)

medianWithoutNA <- function(x) {
   median(x[which(!is.na(x))])
}

k = 6

df[columns_list$outcome_columns] <- lapply(df[columns_list$outcome_columns], as.numeric)

for (column in setdiff(columns_list$numerical_columns, columns_list$outcome_columns)) {
  if (mean(is.na(df[[column]])) > 0.5) next

  df %>%
    mutate(rounded_date_admission_t0 = lubridate::floor_date(date_admission_t0, 
                                                             'month')) %>%
    group_by(rounded_date_admission_t0) %>%
    summarise('Mean' = mean(!!sym(column), na.rm = T),
              'Min' = min(!!sym(column), na.rm = T),
              'Median' = medianWithoutNA(!!sym(column)),
              'Max' = max(!!sym(column), na.rm = T)) %>%
    ungroup %>%
    arrange(rounded_date_admission_t0) %>%
    mutate(Mean = zoo::rollmean(Mean, k = k, fill = NA),
           Min = zoo::rollmean(Min, k = k, fill = NA),
           Median = zoo::rollmean(Median, k = k, fill = NA),
           Max = zoo::rollmean(Max, k = k, fill = NA)) %>%
    pivot_longer(!rounded_date_admission_t0, names_to = 'metric') %>%
    ggplot(aes(x = rounded_date_admission_t0, y = value, color = metric)) + 
      geom_line() +
      geom_vline(xintercept = lubridate::ymd('2020-03-01'), linetype = "dashed") +
      labs(y = column) -> p
      
  print(p)
}
```

```{r warning=F, message=F, fig.show="hold", out.width="50%", fig.height = 5}
df[columns_list$outcome_columns] <- lapply(df[columns_list$outcome_columns], factor)

for (column in c(columns_list$categorical_columns, columns_list$outcome_columns)) {
  if (mean(is.na(df[[column]])) > 0.5) next
  if (length(unique(df[[column]])) > 10) print(column)
  
  df %>%
    mutate(rounded_date_admission_t0 = lubridate::floor_date(date_admission_t0, 
                                                             'month')) %>%
    group_by(rounded_date_admission_t0, !!sym(column)) %>%
    summarise(N = n()) %>%
    ungroup %>%
    arrange(rounded_date_admission_t0) %>%
    mutate(N = zoo::rollmean(N, k = k, fill = NA)) %>%
    group_by(rounded_date_admission_t0) %>%
    mutate(percentage = N / sum(N)) %>%
    ungroup %>%
    tidyr::complete(rounded_date_admission_t0, !!sym(column),
                    fill = list(N = 0, percentage = 0)) %>%
    ggplot(aes(x = rounded_date_admission_t0, y = percentage,
               color = !!sym(column), fill = !!sym(column))) + 
      geom_area(alpha = 0.6) +
      scale_y_continuous(labels = scales::percent) +
      labs(y = column) -> p
      
  print(p)
}
```

```{r}
df %>%
  mutate(rounded_date_admission_t0 = lubridate::floor_date(date_admission_t0, 
                                                           'month')) %>%
  filter(rounded_date_admission_t0 <= lubridate::ymd('2021-08-01'),
         rounded_date_admission_t0 <= lubridate::ymd('2021-01-01')) %>%
  .$aco %>% 
  max(na.rm = T)

column <- 'aco'
df %>%
    mutate(rounded_date_admission_t0 = lubridate::floor_date(date_admission_t0, 
                                                             'month')) %>%
    group_by(rounded_date_admission_t0) %>%
    summarise('Mean' = mean(!!sym(column), na.rm = T),
              'Min' = min(!!sym(column), na.rm = T),
              'Median' = medianWithoutNA(!!sym(column)),
              'Max' = max(!!sym(column), na.rm = T)) %>%
    ungroup %>%
    arrange(rounded_date_admission_t0) %>%
    mutate(Mean = zoo::rollmean(Mean, k = k, fill = NA),
           Min = zoo::rollmean(Min, k = k, fill = NA),
           Median = zoo::rollmean(Median, k = k, fill = NA),
           Max = zoo::rollmean(Max, k = k, fill = NA)) %>%
  filter(rounded_date_admission_t0 <= lubridate::ymd('2021-12-01'),
         rounded_date_admission_t0 >= lubridate::ymd('2021-01-01'))
```
