==> 3-tables.Rmd <==
---
author: "Eduardo Yuki Yada"
geometry: margin=1cm
output: 
  pdf_document:
    template: latex-template.tex
params:
  outcome_column: death_30days
title: "`r paste('Tables -', params$outcome_column)`"
---

```{r, include=FALSE}
knitr::opts_chunk$set(
  warning = F,
  message = F
)
```

# Imports

```{r warning=F, message=F}
library(tidyverse)
library(yaml)
library(kableExtra)
```

# Loading data

```{r}
load('dataset/processed_data.RData') 
load('dataset/processed_dictionary.RData') 

columns_list <- yaml.load_file("./auxiliar/columns_list.yaml")

outcome_column <- params$outcome_column

if (outcome_column == 'general') {
  df <- df %>% mutate(general = 'All')
}

df[columns_list$outcome_columns] <- lapply(df[columns_list$outcome_columns], as.character)
df[columns_list$outcome_columns] <- lapply(df[columns_list$outcome_columns], as.integer)
```

# Numerical variables

```{r results='asis', message=F, warning=F}
medianWithoutNA <- function(x) {
   median(x[which(!is.na(x))])
}
i = 0
for (column in columns_list$numerical_columns) {
  df %>%
    group_by_at(vars(one_of(outcome_column))) %>%
    summarise('Mean' = mean(!!sym(column), na.rm = T),
              'Min' = min(!!sym(column), na.rm = T),
              'Median' = medianWithoutNA(!!sym(column)),
              'Max' = max(!!sym(column), na.rm = T),
              'Standard Deviation' = sd(!!sym(column), na.rm = T),
              'N' = n(),
              'Missing' = sum(is.na(!!sym(column)))) %>%
    ungroup %>%
    mutate(Min = ifelse(is.infinite(Min), NA, Min),
           Max = ifelse(is.infinite(Max), NA, Max)) %>%
    kbl(align = "l", booktabs = T, digits = 3, format = 'latex', label = i, 
        caption = df_names %>% filter(variable.name == column) %>% .$field.label) %>%
    column_spec(1, bold = T, width = "8em") %>%
    row_spec(c(1) - 1, extra_latex_after = "\\rowcolor{gray!6}") %>%
    collapse_rows(1, latex_hline = "none") %>%
    kable_styling(latex_options = c("HOLD_position", "repeat_header")) %>%
    print
  
  i <- i + 1
}
```

# Categorical variables

```{r}
paste_matrix <- function(...,sep = " ", collapse = NULL){
    n <- max(sapply(list(...),nrow))
    p <- max(sapply(list(...),ncol))

    matrix(paste(...,sep = sep,collapse = collapse),n,p)
}
```

```{r}
percent <- function(x) paste0("(", lapply(x, as.character), "%)")

addpercentage <- function(df, horizontal = FALSE) {
  if (horizontal) {
    x <- df %>%
      prop.table(margin = 1) %>%
      addmargins(FUN = list(Total = sum), quiet = TRUE) %>%
      round(2) * 100
    
    x[nrow(x),] <- " " 
    x[-(nrow(x)),] <- lapply(x[-(nrow(x)), ], percent)
  } else {
    x <- df %>%
      prop.table(margin = 2) %>%
      addmargins(FUN = list(Total = sum), quiet = TRUE) %>%
      round(2) * 100
    
    x[, ncol(x)] <- " " 
    x[, -(ncol(x))] <- lapply(x[, -(ncol(x))], percent)
  }
  
  y <- matrix(x, nrow = nrow(df) + 1)
  
  df <- df %>%
    addmargins(FUN = list(Total = sum), quiet = TRUE)
  
  df_final <- paste_matrix(df, y)
  rownames(df_final) <- rownames(df)
  colnames(df_final) <- colnames(df)
  return(df_final)
}
```

```{r results = 'asis', message = F, warning = F}
transpose_columns <- c()
df[columns_list$outcome_columns] <- lapply(df[columns_list$outcome_columns], as.character)

for (column in columns_list$categorical_columns) {
  transpose <- FALSE
  if (length(unique(df[[column]])) > 5) transpose <- TRUE

  variable_name <- df_names %>%
    filter(variable.name == column) %>% 
    .$field.label
  
  abbreviated_name <- df_names %>%
    filter(variable.name == column) %>% 
    .$field.label

  caption <- sprintf('Contingency table between %s and %s', 
                     str_replace(outcome_column, "_", " "),
                     variable_name)

  if (transpose) {
    
    temp_table <- table(df[[column]],
                        df[[outcome_column]],
                        useNA = "ifany") %>%
      addpercentage(horizontal = TRUE) 
    
    has_na <- df[[column]] %>% is.na() %>% sum > 0
    
    if (has_na) {
      rownames(temp_table)[nrow(temp_table) - 1] <- "NA"
    }
    
    t <- temp_table %>%
      as.data.frame %>%
      rownames_to_column(var = abbreviated_name) %>%
      kbl(align = "c", booktabs = T, digits = 2, format = 'latex',
          caption = caption, label = i) %>%
      row_spec(length(unique(df %>% .[[column]] %>% replace_na("NA"))), 
               hline_after = T) %>%
      column_spec(3, border_right = T) %>%
      collapse_rows(1, latex_hline = "none") %>%
      add_header_above(c(setNames(1, ' '),
                         setNames(length(unique(df[[outcome_column]])),
                                  outcome_column))) %>%
      kable_styling(latex_options = c("HOLD_position", "repeat_header"))
    
  } else { 
    temp_table <- table(df[[outcome_column]], 
                        df[[column]],
                        useNA = "ifany") %>%
      addpercentage
    
    has_na <- df[[column]] %>% is.na() %>% sum > 0
    
    if (has_na) {
      colnames(temp_table)[ncol(temp_table) - 1] <- "NA"
    }
    
    t <- temp_table %>%
      as.data.frame %>%
      rownames_to_column(var = outcome_column) %>%
      kbl(align = "c", booktabs = T, digits = 2, format = 'latex',
          caption = caption, label = i) %>%
      row_spec(2, hline_after = T) %>%
      column_spec(length(unique(df %>% .[[column]] %>% replace_na("NA"))) + 1,
                  border_right = T) %>%
      collapse_rows(1, latex_hline = "none") %>%
      add_header_above(c(' ' = 1,
                         setNames(length(unique(df[[column]])),
                                  abbreviated_name))) %>%
      kable_styling(latex_options = c("HOLD_position", "repeat_header")) 
    
  }
  print(t)
  i <- i + 1
}
```
