==> 5-correlations.Rmd <==
---
author: "Eduardo Yuki Yada"
geometry: margin=1cm
output: 
  pdf_document:
    template: latex-template.tex
params:
  outcome_column: readmission_30d
  threshold: 0.1
title: "`r paste('Correlations -', params$outcome_column)`"
---

```{r, include=FALSE}
knitr::opts_chunk$set(
  warning = F,
  message = F
)
```

# Imports

```{r warning=F, message=F}
library(tidyverse)
library(yaml)
library(ggcorrplot)
library(kableExtra)

source("aux_functions.R")
```

# Loading data

```{r}
load('dataset/processed_data.RData') 
load('dataset/processed_dictionary.RData') 

columns_list <- yaml.load_file("./auxiliar/columns_list.yaml")

outcome_column <- params$outcome_column
threshold <- params$threshold
print(threshold)

df[columns_list$outcome_columns] <- lapply(df[columns_list$outcome_columns], as.character)
df[columns_list$outcome_columns] <- lapply(df[columns_list$outcome_columns], as.integer)
```

# Correlation

```{r}
na_eligible_columns <- df %>% 
  summarise(across(everything(), ~ mean(is.na(.)))) %>%
  select_if(function(.) last(.) < 0.9) %>%
  names

unique_eligible_columns <- df %>% 
  summarise(across(everything(), ~ length(unique(.)))) %>%
  select_if(function(.) last(.) > 1) %>%
  names

t0_columns = df_names %>%
  filter(momento.aquisicao == 'Admissão t0') %>%
  .$variable.name

exception_columns <- c('dieta_enteral')

eligible_columns <- intersect(na_eligible_columns,
                              unique_eligible_columns) %>%
  intersect(t0_columns) %>%
  setdiff(exception_columns)

corr <- df %>% 
  select(all_of(intersect(columns_list$numerical_columns,
                          eligible_columns))) %>%
  cor(use = "pairwise.complete.obs") %>%
  as.matrix

corr_table <- corr %>%
  as.data.frame %>%
  tibble::rownames_to_column(var = 'row') %>% 
  tidyr::pivot_longer(-row, names_to = 'column', values_to = 'correlation')

rename_column <- function(df, column_name){
  variable.name <- 'variable.name'
  df <- df %>%
    left_join(df_names %>% select(variable.name, abbrev.field.label),
              by = setNames(variable.name, column_name)) %>%
    select(-all_of(column_name)) %>%
    rename(!!sym(column_name) := abbrev.field.label) %>%
    relocate(!!sym(column_name))
}

corr_table %>%  
  filter(correlation > 0.8) %>%
  filter(row < column) %>%
  rename_column('row') %>%
  rename_column('column') %>%
  select(row, column, correlation) %>%
  niceFormatting(caption = "Pearson Correlation", font_size = 9, label = 1)
```
```{r echo=F, results='asis'}
cols <- c('antiarritmico', 'betabloqueador', 'bloq_calcio', 'dva', 'espironolactona', 'ieca_bra', 'insuf_cardiaca')

corr_table %>%
  filter(row %in% cols, column == 'meds_cardiovasc_qtde') %>%
  rename_column('row') %>%
  rename_column('column') %>%
  niceFormatting(caption = "meds cardiovasc qtde correlations", 
                 font_size = 9, label = 2)

cols <- c('radiografia')

corr_table %>%
  filter(row %in% cols, column == 'exames_imagem_qtde') %>%
  rename_column('row') %>%
  rename_column('column') %>%
  niceFormatting(caption = "exames imagem qtde correlations", 
                 font_size = 9, label = 3)

cols <- c('ecg')

corr_table %>%
  filter(row %in% cols, column == 'metodos_graficos_qtde') %>%
  rename_column('row') %>%
  rename_column('column') %>%
  niceFormatting(caption = "metodos graficos qtde correlations", 
                 font_size = 9, label = 4)
```

# Hypothesis Tests

```{r}
df_wilcox <- tibble()

for (variable in intersect(columns_list$numerical_columns,
                           eligible_columns)) {
  if (mean(is.na(df[[variable]])) > 0.95) next
  
  x <- filter(df, !!sym(outcome_column) == 0)[[variable]]
  y <- filter(df, !!sym(outcome_column) == 1)[[variable]]
  
  test = tryCatch(wilcox.test(x, y, alternative = "two.sided", exact = FALSE), 
                  error = function(cond) {
                    message("Can't calculate Wilcox test for variable ", variable)
                    message(cond)
                    return(list(statistic = NaN, p.value = NaN))
                  })
  
  df_wilcox = bind_rows(df_wilcox,
                        list("Variable" = variable, 
                             "Statistic" = test$statistic,
                             "p-value" = test$p.value))
}

significant_num_cols <- df_wilcox %>%
  filter(`p-value` <= threshold) %>%
  select(Variable) %>%
  pull

df_wilcox <- df_wilcox %>%
  arrange(`p-value`) %>%
  mutate(`Statistic`  = round(`Statistic`, 3)) %>%
  rename_column('Variable')

df_wilcox %>%
  mutate(`p-value` = case_when(`p-value` == 1 ~ sprintf('> 0%s999', getOption("OutDec")), 
                               `p-value` < 0.001 ~ sprintf('< 0%s001', getOption("OutDec")), 
                               TRUE ~ as.character(round(`p-value`, 3)))) %>%
  niceFormatting(caption = "Mann-Whitney Test", label = 5)
```

```{r}
df_chisq <- tibble()

for (variable in intersect(columns_list$categorical_columns,
                           eligible_columns)) {
  if (length(unique(df[[variable]])) > 1) {
    test <- tryCatch(chisq.test(df[[outcome_column]], 
                     df[[variable]] %>% replace_na('NA'), # counting NA as cat
                     simulate.p.value = TRUE),
                     error = function(cond) {
                       message("Can't calculate Chi Squared test for variable ", variable)
                       message(cond)
                       return(list(statistic = NaN, p.value = NaN))
                     })

    df_chisq <- bind_rows(df_chisq,
                         list("Variable" = variable,
                              "Statistic" = test$statistic, 
                              "p-value" = test$p.value))
  }
}

significant_cat_cols <- df_chisq %>%
  filter(`p-value` <= threshold) %>%
  select(Variable) %>%
  pull

df_chisq <- df_chisq %>%
  arrange(`p-value`) %>%
  mutate(`Statistic`  = round(`Statistic`, 3)) %>%
  rename_column('Variable')

df_chisq %>%
  mutate(`p-value` = case_when(`p-value` == 1 ~ sprintf('> 0%s999', getOption("OutDec")), 
                               `p-value` < 0.001 ~ sprintf('< 0%s001', getOption("OutDec")), 
                               TRUE ~ as.character(round(`p-value`, 3)))) %>%
  niceFormatting(caption = "Chi-squared test", label = 6)
```

```{r}
dir.create(file.path("./auxiliar/significant_columns/"), showWarnings = FALSE)

con <- file(sprintf("./auxiliar/significant_columns/categorical_%s.yaml", outcome_column), "w")
write_yaml(significant_cat_cols, con)
close(con)

con <- file(sprintf("./auxiliar/significant_columns/numerical_%s.yaml", outcome_column), "w")
write_yaml(significant_num_cols, con)
close(con)
```

```{r echo=F, results==F}
print(length(columns_list$categorical_columns))
print(length(significant_cat_cols))
print(length(columns_list$numerical_columns))
print(length(significant_num_cols))

cat(paste(shQuote(significant_cat_cols, type = "cmd"), collapse = ", "))
cat(paste(shQuote(significant_num_cols, type = "cmd"), collapse = ", "))
```
